Move Code that optimizes Imports into its own class and register it
This allows the user to press "Code" -> "Optimize Imports" or the corresponding Keyboard Shortcut to optimize the Imports of the current file
