+++ src/main/resources/META-INF/plugin.xml
@@ -194,6 +194,8 @@
         <lang.smartEnterProcessor language="Elm" implementationClass="org.elm.ide.typing.ElmSmartEnterProcessor"/>
         <extendWordSelectionHandler implementation="org.elm.ide.wordSelection.ElmDeclAnnotationSelectionHandler"/>
 
+        <lang.importOptimizer language="Elm" implementationClass="org.elm.ide.code.format.ElmImportOptimizer" />
+
         <!-- Inspections -->
 
         <localInspection language="Elm" groupName="Elm"
